Layer1-Curriculum/day-02-structured-outputs/requirements.txt
openai
python-dotenv
tiktoken
pydantic
streamlit
colorama
